backend/requirement.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python
numpy
torch
torchvision
scikit-learn
requests
imagehash
python-dotenv
